Move time fields to top-level in bundle detail sidebar.
